text and link update: Updated the text on ArgentX (both about and direct link to access on chrome store)
I updated the text of `ArgentX being the one and only popular wallet` to it being `one of the popular` as for now there is also another called `Bravos` which is also quite popular in the starknet ecosystem. 

Also, the link to the Chrome store doesn't take one directly to the ArgentX wallet, I was also able to fix that.

Thank you!